+++ src/core/marching_cubes.cpp
@@ -173,7 +173,7 @@ void computeTriangles(DomainInfo domain, float3 resolution,
         }
     }
 
-    for (int i = 0; i < indices.size(); i += 3) {
+    for (size_t i = 0; i < indices.size(); i += 3) {
         Triangle t;
         t.a = domain.global2local(vertices[indices[i+0]]);
         t.b = domain.global2local(vertices[indices[i+1]]);
